requirements.txt
paho-mqtt>=2.0.0
supabase>=2.4.0
python-dotenv>=1.0.0
streamlit>=1.32.0
plotly>=5.20.0
pandas>=2.2.0
